iwlwifi: support debugging constant overrides
This is mostly for MVM right now, but in general it supports
other overrides as well.

The general approach is to define the constants for upstream
as constants, but make them macros in a way that they can be
overridden easily under #ifdef.

The values are single files under iwlwifi debugfs, anything
else can be handled in userspace.

When a new device is bound, a snapshot of the current config
is stored in the transport struct (it belongs to the device)
and can then be used by anything else on top of it.

Defaults are provided by including the mvm/constants.h file
with #undef, so they only need to be written into the code a
single time.

Obviously, mvm/constants.h should only contain values in the
non-ifdef'ed portion that can be upstreamed, DBGM config and
similar should go into another file, probably just directly
the iwl-dbg-cfg.h file.

Change-Id: I36ae69a57957b2be13262b6f7fee37653564c441
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Amnon Paz <amnonx.paz@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/16451
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 3294fc968ba01f4ab449dd610bc43d192247d917