iwlwifi: mvm: BT Coex - use data from firmware
The data in MailBox comes direclty from the BT core.
We should use the data processed by the WiFi fw that is
appended to the MailBox in the BT Coex notification.
Also decide on whether the Coex type based on the input
from the the firmware and not hard coded.

Also fix the SMPS SISO threshold to 2 (it was 3).

Change-Id: I1d3fff5c3ecd7b1206432f0d59729c75c1df2be3
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/21693
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: dbdd03ca218febd1f80242a4cf1c08918ecf4b40