iwlwifi: mvm: remove extra SSID from probe request [BUGFIX]
Bits 1-21 in this channel type attributes are indication for which SSID
is going to be sent on this channel. Since the first SSID is sent implicitly
in the probe request, we don't need to toggle it's bit here.

Change-Id: I694ebee1c13848809414cbe907d26802a5fe2c9e
MWG100181355
Signed-off-by: David Spinadel <david.spinadel@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/17703
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: abd069c7f7d58cf2328c886d3dbaa48fc7c74f4c