iwlwifi: mvm: fix D3 image configuration
In my commit "iwlwifi: mvm: support six IPv6 addresses in D3"
I accidentally changed both the old and the new API, when of
course only the new API should be introduced and the old only
be adjusted for the structure repacking. This made the code
send two bytes too many to the device which then lead to the
firmware crashing on the *next* command (because it processes
everything in multiples of 4 bytes only.) Restore the correct
struct to fix this.

Change-Id: Iff997c915dd249d2df8b09e34e0c91b8b1cefc15
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/19557
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 418567d73801e7743644582055001c4131ae4fbb