iwlwifi: mvm: add support for RF-kill on suspend
This is necessary for a customer who calls suspend without
actually suspending the system, if we don't fake rf-kill
then wpa_supplicant would try to access a dead device. Add
this as an option to our code.

Note that this breaks WoWLAN/D3, so if the Kconfig option
is enabled it isn't available.

Change-Id: I632d8491cf5cf8120b2ce5eb533b05d1c0bce676
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/20780

x-iwlwifi-stack-dev: 096daadc662107bd9a3e4b5223ddcece573ea747