iwlwifi: integrate suspend/resume into the driver
Suspend/resume happens in the stack, but that's causing problems
with some special behaviour for a customer. Call the higher-level
stack from within the driver to not have to rely on the kernel
framework that they're ignoring and bypassing.

Note that by itself this patch doesn't really do anything, it's
just there to not always do this. It is safe to always do this,
but there's no justification for upstreaming it and changing all
of the upstream drivers.

Change-Id: Icd18b28ce73caf2af2a5c26583a536701e03fbe9
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/20779

x-iwlwifi-stack-dev: 4c6bd73a6887a00639df61a52eb3477ae2c5bc11