iwlwifi: mvm: add pm_generic_suspend/resume test debugfs file
This is pretty odd, but there's a customer doing it and to
test it this debugfs file is useful. Make it depend on the
CPTCFG_IWLWIFI_MVM_RFKILL_ON_SUSPEND config option to not
mess up the system when the file is used otherwise.

Change-Id: I2bde61981ad515e05b86fd47d67ce2853d3b427a
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/20719
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 666f2ec908ad4362d32992143d5c0a53612c31f2