iwlwifi: mvm: debugfs: add an option to set antennas for scan command
Add an option to set rx antennas for the scan command from debugfs.
Create a file called ant_rxchain in the mvm debugfs directory.
To choose antennas, write a number between 1-7 to ant_rxchain.
Write 1 for A, 2 for B, 3 for AB and so on.

Change-Id: I2d1a6304d8b90ef833a6709ada1f59c9b8ee3ca2
Signed-off-by: Oren Givon <oren.givon@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/20923
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: fbeb7f0a945a6071d8dd41757db2b2ab899a851a