iwlwifi: add transport tester[RFC]
Load driver in transport-tester operational mode.
In this mode the transport layer can be tested separately
from the rest of the stack, and irrespectively of the bus type
itself. using a character device, the user can directly
test the loopback between the transport layer and the fw, while
performing tx, rx and hcmd sending.

This special operational mode requires the adjasted fw file, and
will not be able to run without it. Therefore, the trans_test
flag is disabled in this patch and mvm is loaded.

I am adding the scripts that run this tester for the sake of the
code review only.

Change-Id: I8c646c2de0d659554797dfdbb24bc3e14363e856
Signed-off-by: Lilach Edelstein <lilach.edelstein@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/21917
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 437cc6926190f6760ad1a02c788ee74b24e1de98