Revert "iwlwifi: mvm: use IEEE80211_SMPS_OFF by default"
This reverts commit 6ca45374731d93a77eebb15b7915bcdbea947e63.

The problem that commit tried to solve seems to be a bug in the AP
(Linksys EA6400) when using 11ac and SMPS.  With lower rates (eg. 11n
@ 40MHz) the degradation is much smaller and, at least with one other
AP (TP-Link Archer C7), the degradation with 11ac and SMPS is just
about 3.2%.

The reason for slow throughput when Dynamic SMPS is enabled with
Linksys EA6400 should be investigated.  But the feature should not
just be disabled outright.

Change-Id: I8e8549259124d4fc2e9c355f72d4be5ff711ae2c
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/22600
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 6cee1ab4ede00711fce231277d1c914a5f8f0929