iwlwifi: mvm: call ieee80211_scan_completed when needed [BUGFIX]
When RFKill cuts short a scan, mac80211 cancels the scan.
This is done by sending a host command to the firmware, but
this command was dropped because of RFKill. Flag this
command as "SEND_IN_RFKILL" to make sure it is sent to the
firmware. The firmware will send SCAN_COMPLETE_NOTIFICATION
which will trigger a call to ieee80211_scan_completed.

If the scan cannot be aborted, it is because the firmware
already finished the scan but we hadn't notified mac80211
at the time mac80211 decided to cancel the scan. By the time
we see the scan could not be aborted, mac80211 has been
notified already.

This patch fixes situations in which we didn't notify
mac80211 upon completion of the scan that was cut short
by RFkill.

Cc: stable@vger.kernel.org [3.10+]
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
---
Change-Id: I1523f76adc89aaf43f0a8fa3face3421310476bc
MWG100186316
Reviewed-on: https://gerrit.rds.intel.com/22560
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: e6cdca375d68805e29e6d85d5dbc6d408393b7ca