Revert "mac80211: allow configuring smps_mode_in_ps"
This reverts commit fe99884b7d65143b4b04ae22400a0c19040501c3.

We probably don't need to disable SMPS by default, since the reason
for doing it seems to be a bug in an access point.  Thus, we don't
need this patch in mac80211 either.  This patch can be added back if
needed again in the future.

Change-Id: Ibe05daf3727ef333ac16983e28a8efe791e864fe
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/22601
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 0527da1b4c12ff5421cf8050254bcac9a63dcc7e