iwlwifi: mvm: BT Coex remove a lockdep warning
We need to compute the AMPDU size limit from rate control
code which can't take mvm->mutex. This means that there is
a race but in the worst case, we will have a wrong AMPDU
size limit which is not a big issue.

Change-Id: I167ffd5f030ea6b89e7ad7c05d687596488eaba6
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/23594
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 550f5427a5398e454a11d83355c991b8317767bd