iwlwifi: mvm: Add Smart FIFO support
Send firmware a Smart FIFO Configuration host command to allow
interrupt coalescing. The smart FIFO is enabled when there is
only one bound interface (other than  p2p devices which are
ignored) and it is of type station, and activated while the
station is associated.

Smart Fifo allows aggragations of DMA transactions and by that
causes processor and memory controller to stay for a longer time
on lower c-states, thus saving platform power.
Firmware relies on driver to activate and disable it.

Change-Id: I86a2a11f87a4434fd6b96ebfcdab1325fc5047d0
Signed-off-by: lilach edelstein <lilach.edelstein@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/23413
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: ba23773ee903ce9eb22268363248159a1423df18