iwlwifi: pcie: poke device when commands don't complete quickly [BUGFIX]
In certain corner cases in the firmware implementation, powersave
transitions can cause the firmware to miss the fact that commands
were added to the queue/FIFO and thus never processes them. Since
the commands really are in the queue, try to poke the firmware in
such cases (by grabbing NIC access, which wakes up the NIC) so it
notices the new command and processes it.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
---

MWG100185613
MWG100185616

Change-Id: I9621a0a05a7af9398cd936891425f03c91ce8a33
Reviewed-on: https://gerrit.rds.intel.com/23825
Tested-by: IWL Jenkins
Reviewed-by: Alexander Bondar <alexander.bondar@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 2bf517aa8da8d9c49eddc21e2adbb0bbe1363f91