cfg80211: fix scheduled scan pointer access
Since rdev->sched_scan_req is dereferenced outside the
lock protecting it, this might be done at the wrong
time, causing crashes. Move the dereference to where
it should be - inside the RTNL locked section.

Cc: stable@vger.kernel.org [3.8+]
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
---
This *could* be the cause of the crash, it was the only
thing I saw ...

Change-Id: I9e0c2ae52639c7ca3753ee734e3e2486de9deb22
Reviewed-on: https://gerrit.rds.intel.com/24451
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: b7ff730f574bbd8d9a0009311cabfef990f44ffc