iwlwifi: remove TX_CMD id from transport layer
The transport layer doesn't need to know the TX_CMD id.
It can be set by the op_mode.
The transport layer still needs to know the layout of the
Tx command because of alignment issues and because of the
scratch pointer.

Change-Id: Icd9d6abf34d49a0eeba8af9b118d80f1410dfdc6
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/24529
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 92eb8c58cb5121b3037ba3c4631fa365b9a708cd