iwlwifi: mvm: clear the binding before restart
mac80211 will re-assign the channel to the vif anyway.
Not doing so can lead to an endless SYSASSERT loop as
reported by Catalin:
If the binding command happens to fail, then mac80211
will consider this vif as un-bound which means that
ieee80211_assign_chanctx will not do anything when called
from the reconfig flow.
In this case, the call to update quota after the restart is
complete will make the firmware unhappy since we will add
quota to a vif that is not bound.

Fix this by setting the vif as not bound before the restart
which will hint update_quota that no quota should be set
for this vif.

Change-Id: If6d7700ee47a31ce231bbc9cbeeee0a9127f9b66
Cc: stable@vger.kernel.org [3.10+]
Fixes: 8ca151b568b6 ("iwlwifi: add the MVM driver")
Reported-by: Catalin Drula <catalin@suitabletech.com>
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/25066
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 127ca3a32e720b3a2f706178e10b845698d88ebd