iwlwifi: mvm: fix ht protection flags [BUGFIX]
The HT protection flags have been a mess for a long time,
this patch fixes all this.

The proper source of information for this is the
protection in the HT operation mode IE which is
propagated to bss_conf by mac80211.

1) No need to set the HT protection flags when the link
   doesn't support HT.
2) We need to set the TGG protection when ERP is active.

Change-Id: Id2ac63ffd41f8155e3a0295500cf5c639284329c
Signed-off-by: Avri Altman <avri.altman@intel.com>
---
MWG100183228
Reviewed-on: https://gerrit.rds.intel.com/25253
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 92cb3264948e1ad0fa93afec352be90a6ac218bf