iwlwifi: mvm: rs: update expected TPT tables if aggregation changed
Expected TPT table was updated only when switching to a new modulation.
This is wrong as aggregation being turn on or off changes the expected TPT
signficantly. This lead to scenarios where aggregation turned on after being
in MIMO lead to switching back to SISO despite a perfect success ratio.
The reason being that the TPT of the SISO mode was being estimated based
on aggregation while the MIMO one wasn't.

Change-Id: I5041a2b2477decd156c230e0dde155e8de3f15bf
Signed-off-by: Eyal Shapira <eyal@wizery.com>
Reviewed-on: https://gerrit.rds.intel.com/26750
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 13cda0ed7023f0f0a02ddb0dae8a83771d8fc8c6