iwlwifi: mvm: stop_device when the module is loaded in RFKILL
This allows to disable ICT and load the firmware when
RFKILL will be deasserted. ICT interrupts can't be used
to load the firmware and not calling to stop_device left
ICT enabled, hence the bug.
Also move the call to stop_device out of run_init_mvm_ucode.
Whie at it, WARN if we call run_init_ucode when the
init_ucode_complete flag is raised - this would be a bug.

Change-Id: I1304008bc696a3234da61fb9b45e511cb3e681be
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/27295
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 1992c093568b54fa5a4af4b44bc512ce37419e04