iwlwifi: pcie: track interrupt mask in SW
Track the interrupt mask in software, making it exactly
what is configured in the interrupt mask register in the
hardware.
This allows not to access the register from the interrupt
handler. This was the case for ICT interrupt already, but
not for non-ICT interrupt.

Change-Id: I28748fe7c896afc251a6b077bc4504df7008120a
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/27317
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: a89af25d31110162dc67efa5e75f44e8b1fda5da