iwlwifi: mvm: set highest rate in VHT MCS Set
Set the Tx/Rx highest long GI rates in the VHT Supported MCS Set
field according to the chip capabilties.

Change-Id: I7d38154141d7b22f0ecd066a3a324c331f0c7650
Signed-off-by: Eyal Shapira <eyal@wizery.com>
---
This was reported as a bug during 11ac integration as it doesn't
match the Connectivity Management SAS requirement and the Windows
driver behaviour which is the reference here.
The value 0 was ok AFAICT according to the spec section 9.7.11.1
but it might be good to set this anyways.
Reviewed-on: https://gerrit.rds.intel.com/27602
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 812ad97bba8f2246e8a9f7ec56d15d099f6e280e