iwlwifi: mvm: fix smart FIFO locking/station pointer use
The smart FIFO code can be called from atomic context from the
sta_rc_update() callback, when a VHT peer's NSS changes. This
means that it can't rely on having the mvm->mutex held, which
makes it unsafe to take a vif pointer from iteration and use
it outside the iteration, as well as making it impossible to
use rcu_dereference_protected() that relies on the mutex being
held.

Fix the first issue by gathering only the data we need in the
iterator code, and the second by using proper RCU protection.

Change-Id: Iccd6d5338f570400693b0783c4ba985352b14443
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/27844
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 54daa926d0c06b71a88c2a867081af68599aef34