iwlwifi: mvm: disable the device as soon as RFKILL fires
The firmware needs to be stopped quickly (100ms) after the
RFKILL interrupt fired. Failing to do so would allow the
firmware to access the radio registers which would lead to
a hardware error.

Before this change, we would kill the firmware only when
mac80211 stops the device which can take a fair amount of
time. Take a shortcut by stopping the device right away
in the interrupt.

This is not relevant if the current firwamre is INIT
firmware since that firmware can run while in RFKILL.

Change-Id: I000f9a297c081b3d37acb8987a0409d001a6e067
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/27980

x-iwlwifi-stack-dev: da9ef856a4c4e3811915d8f4cf81cfab843e3550