iwlwifi: testmode: fix alignment of netlink message structs
Running the driver on a 64-bit machine changes the sizes
of the structs used to communicate between the iwlwifi
driver and the xVT service USC, due to an alignment problem.

As a quick fix to still allow communication until the xVT
Service structs are all fixed to allow compilation for a
64bit machine, this patch makes sure the alignment of a
32bit compiled xVT Service that runs on a 64bit machine is
compatible with a 64bit compiled driver. The structs that
hold the messages on the driver side are packed and 4-byte
aligned, thus maintaining the same alignment as the
original 32bit compiled xVT and allowing netlink
communication.

Change-Id: I8e5d218a8791c89ab9a6476e4bf6ab4e6000e325
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/28344
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 583730a40d893b255b637cc4adb3cca9dba7d78d