iwlwifi: mvm: don't use highest rate in VHT MCS Set
Keeping this as 0 is ok according to spec section 9.7.11
as this means the limits are according to the Tx/Rx
supported MCS x NSS bitmap. Initially we've set these as
there were concerns of interop issues but these turned out
to be false.

Change-Id: I60040441340e9e2541e6c2a28d4ba92217ff660b
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
---
This is a revert of https://gerrit.rds.intel.com/#/c/27602/
The requirements to set these values were getting more complex
and required setting this as X in probe requests according to our
cap and Y in the assoc request according to the peer supported values
for interop reasons.
After discussing more with system (Ayelet) and checking if there were
any real interop issues reported we agreed there's no real need to
set these and we can keep using 0 which is ok by the spec.
The original patch was upstreamed already so we can't squash.
Reviewed-on: https://gerrit.rds.intel.com/28497
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 40af8584f46e32ef69e8cf8b9ac9809e37c1a405